requirements.txt
# Production dependencies
playwright>=1.40.0
pyyaml>=6.0
openai>=1.0.0
httpx[socks]>=0.28.0  # SOCKS proxy support for DeepSeek API client (optional but recommended if using proxy)

